% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_param.R
\name{check_param}
\alias{check_param}
\title{Check params}
\usage{
check_param(param, allowed_params)
}
\arguments{
\item{param}{A character will be used as parameter in a function.}

\item{allowed_params}{Allowed characters used as parameter in a function.}
}
\value{
Zero value if the param match the allowed_params, 1 if input are invalid, 2 if param does not match.
}
\description{
Retrieve KEGG pathway names for a specific organism.
}
