% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GETKEGG_Module.R
\name{GETKEGG_Module}
\alias{GETKEGG_Module}
\title{Specific function to retrive KEGG Modules names using local database and https://bio2rdf.org/ resurces}
\usage{
GETKEGG_Module(modules_IDs)
}
\arguments{
\item{modules_IDs}{A KEGG code string (e.g., modules_IDs <- c("M00005", "M00045", "M00060")).}
}
\value{
A data frame with KEGG modules IDs and names.
}
\description{
Retrieve KEGG modules IDs.
}
