% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GETKEGGNames.R
\name{GETKEGGNames}
\alias{GETKEGGNames}
\title{Get KEGG names using https://bio2rdf.org/ resurces}
\usage{
GETKEGGNames(KEGG_IDs)
}
\arguments{
\item{KEGG_IDs}{A KEGG code string (e.g., KEGG_IDs <- c("map01100", "map01110", "map01120")).}
}
\value{
A data frame with IDs and names.
}
\description{
Retrieve KEGG ID names for a specific organism.
}
