% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGList.R
\name{KEGGList}
\alias{KEGGList}
\title{Get KEGG names using KEGG API}
\usage{
KEGGList(entry, database)
}
\arguments{
\item{entry}{A KEGG code string (e.g., entry <- c("map01100", "map01110", "map01120")).}

\item{database}{Define selected KEGG database to query.}
}
\value{
A data frame with IDs and names.
}
\description{
Retrieve KEGG pathway names for a specific organism.
}
